Major smart contract updates (#23)
* Add vesting, syndication fee, DealDex fee

* Add support to associate investments with NFT ID

By default, investments are still associated with investor wallet

* Change investment token from ETH to arbitrary ERC20

I do not enforce USDC or USDT in the smart contract. I'm not convinced
we should force that - and if we do, that belongs in the frontend code.

* Add exchange rate and refund overrides

I do not use FixedPoint. It adds nothing beyond saving 256 bits of
storage

* Update setProjectToken interface to allow updating exchange rate

Co-authored-by: Kapil Sinha <ksinha1999@gmail.com>